backend-api/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg[binary]==3.2.1
celery==5.4.0
redis==5.0.7
requests==2.32.3
polyline==2.0.0
gunicorn==22.0.0